requirements.txt
numpy==1.18.1
opencv_python==4.2.0.32
Pillow==7.1.1
